(refactor) Rename server routes to controllers
