+++ .gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 .AppleDouble
 .LSOverride
+.env
 azure.err
 iisnode.yml
 coverage
