+++ .sample-env
@@ -1,0 +1,3 @@
+FACEBOOK_ID='facebook id'
+FACEBOOK_SECRET='facebook secret'
+GEO_API_KEY='google geocode API key'
