+++ package.json
@@ -11,6 +11,7 @@
   },
   "scripts": {
     "test": "mocha --reporter=landing",
+    "setup": "node setup.js",
     "start": "./node_modules/.bin/nodemon ./server/server.js --ignore client/",
     "cover": "./node_modules/.bin/babel-node ./node_modules/istanbul/lib/cli cover _mocha -- --bail --recursive"
   },
@@ -24,8 +25,10 @@
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "body-parser": "^1.15.2",
+    "dotenv": "^2.0.0",
     "express": "^4.14.0",
     "express-session": "^1.14.0",
+    "fs": "0.0.2",
     "knex": "^0.11.9",
     "nodemon": "^1.10.0",
     "passport": "^0.3.2",
