+++ server/passport.js
@@ -1,8 +1,9 @@
 var FacebookStrategy = require('passport-facebook').Strategy;
-const FACEBOOK_ID = '639261402906273'
-const FACEBOOK_SECRET = 'fdf4dc317a7332195fb5b2278b522e68'
+const FACEBOOK_ID = process.env.FACEBOOK_ID;
+const FACEBOOK_SECRET = process.env.FACEBOOK_SECRET;
 const FACEBOOK_CALLBACK_URL = 'http://localhost:3000/api/passportFacebook/facebookLogin/Callback'
 const PROFILE_FIELDS = ['id', 'email', 'gender', 'link', 'locale', 'name', 'timezone', 'updated_time', 'picture.type(large)', 'verified', 'friends'] //This is what tells facebook what to return		
+
 var User = require('./models/user');
 var PassportHelper = require('./passport_helper')
