+++ setup.js
@@ -1,0 +1,4 @@
+'use strict';
+var fs = require('fs');
+fs.createReadStream('.sample-env')
+  .pipe(fs.createWriteStream('.env-test'));
